Move setup code into fixture; avoid global; add JSON.stringify, __proto__
